agentura-ui/pyproject.toml
[project]
name = "agentura-ui"
version = "0.1.0"
description = "Panel-based chat client for Semos Agentura MCP agents"
requires-python = ">=3.13"
dependencies = [
    "panelini[ai]",
    "litellm",
    "langchain-community",
    "mcp>=1.9",
    "httpx>=0.28",
    "python-dotenv>=1.0",
]

[tool.uv.sources]
panelini = { git = "https://github.com/opensemanticworld/panelini.git", branch = "ai-component" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentura_ui"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
]
ui = [
    "pytest-playwright>=0.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["ui: browser tests requiring Playwright"]
